backend-fastapi/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase
supabase>=2.4.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OAuth
authlib==1.3.0
httpx>=0.24,<0.26

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Validation
email-validator==2.1.0
